// src/app/BusinessModel/IceHockey/IceHockeyMatchStatusEnum.enum.ts
export enum IceHockeyMatchStatusEnum {
    NSY = 57,
    FINISHED = 58,
    FIRST_PERIOD = 59,
    SECOND_PERIOD = 60,
    THIRD_PERIOD = 61,
    BREAK = 62,
    OVER_TIME = 63,
    FINISHED_OT = 109,
    FINISHED_AP = 110,
    PENALITY = 111,
    POSTPONED = 118,
    AWARDED = 126,
    RESULT_ONLY = 133,
    CANCELLED = 191,
    INTERRUPTED = 197,
    DELAYED = 198,
    NO_INFO_YET = 205,
    ABANDONED = 219
}