// src/app/BusinessModel/Basketball/basketball-match-status-enum.ts
export enum BasketBallMatchStatusEnum {
    FIRST_QUARTER = 88,
    SECOND_QUARTER = 89,
    THIRD_QUARTER = 90,
    FORTH_QUARTER = 91,
    NSY = 93,
    FINISHED = 94,
    FINISHED_OT = 112,
    OVER_TIME = 113,
    BREAK = 92,
    POSTPONED = 119,
    RESULT_ONLY = 128,
    AWARDED = 124,
    CANCELLED = 134,
    DELAYED = 199,
    NO_INFO_YET = 209,
    INTERRUPTED = 210,
    ABANDONED = 218,
}